going back to working version old project
reverted to working version from the old project from the previous
students
https://git.fhict.nl/I275684/Fontys-EGT/commit/e9efa664c85e57a183c6d69c7511b2052bb226be
then i added the things they added to make the commit after this commit
and it worked. this was deleting the toolbar so that the settings menu
would pop out, and so that the about section doesnt exist anymore. ill
have to add everything they did step by step now to get back to where
they were so that i can move on.