fixed the problem. back with the latest version from former students gitlab
the problem was that the former student working on this project was
making changes to the GUI by deleting things he didnt need. he then went
on to delete the code that had to de with these GUI components EXCEPT
the signals/slots in the header files.. which resulted in to an error
refering to the MOC files. in these moc files there were pointers
refering to the deleted GUI components. so instead of deleting the
signals/slots in the header files, the former student deleted the
pointers in the moc file. which resulted in a build that worked for
him... but obviously when i got the source code. and compiled it. these
errors came back. and since i wasnt familiar with what he had done to
get these errors. it took me a while to figure out what the exect
problem was. but now its fixed and i can move on and start on my own
project