JSON Converter and WebSocket
Added the JSON converter to convert the array of structs containing the
tag data to JSON. this JSON data is then send to the webserver using
websockets. i used the QtWebsockets to make my client. theres no real
error checking yet. so if the server were te be down and that code was
asked to run it could result in to an error at this point. but the
server runs 24/7 most of the time. so for now this will do